Finish settings refactor - all imported linearly
