+++ settings/_env.py
@@ -10,7 +10,7 @@ assert ENV in ('dev', 'test', 'prod')
 DEBUG = bool(environ.get('DEBUG'))
 assert not (DEBUG and ENV == 'prod'), "DEBUG can't be set in prod"
 if DEBUG and ENV == 'test':
-    logger.warning('Ignoring DEBUG value in test env: setting DEBUG to False')
+    logger.warning('DEBUG value is ignored in test env and always set to False')
     DEBUG = False
 
 SECRET_KEY = environ.get('PP_SECRET_KEY') or environ.get('SECRET_KEY')
