+++ settings/celery.py
